+++ gneiss-mqtt/src/error.rs
@@ -509,6 +509,7 @@ impl From<tokio::sync::oneshot::error::RecvError> for MqttError {
     }
 }
 
+
 impl <T> From<std::sync::mpsc::SendError<T>> for MqttError where T : Send + Sync + 'static {
     fn from(err: std::sync::mpsc::SendError<T>) -> Self {
         MqttError::new_operation_channel_failure(err)